server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openai==1.3.7
langchain==0.1.0
pypdf==3.17.4
python-dotenv==1.0.0
langgraph==0.0.20
langchain-openai==0.0.5</content>
<parameter name="filePath">v:/ai_interview/server/requirements.txt